BrainSprite Image Load and IntroJS fixes
(1) Using the imagesLoaded JS lib from
https://imagesloaded.desandro.com/
to ensure that rest of JS is not run until the PNGs are loaded

(2) Fixing the second step of the intro so that the tool tip describing
the tree view is now visible (pushed 200px downward). This is done by
watching the position of the tooltip and changing it after IntroJS
positions it.